work with annotations

+++ src/Controller/Bonus/DeleteBonusController.php
@@ -2,8 +2,8 @@
 
 namespace App\Controller\Bonus;
 
-use App\Entity\Bonus\Bonus;
 use App\Manager\BaseManager;
+use App\Repository\Bonus\BonusRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
@@ -11,23 +11,37 @@ use OpenApi\Attributes as OA;
 
 class DeleteBonusController extends AbstractController
 {
-    #[Route('/api/bonus/{bonus}', methods: ['DELETE'])]
-//    #[OA\Parameter(
-//        description: 'Bonus ID',
-//        name: 'ID',
-//        required: true,
-//        in: 'path',
-//        schema: new OA\Schema(type: 'int')
-//    )]
+    #[Route('/api/bonus/{bonusId}', requirements: ['bonusId' => '\d+'], methods: ['DELETE'])]
     #[OA\Response(
-        response: 200,
+        response: Response::HTTP_OK,
         description: 'Successfully action',
+        content: new OA\JsonContent(
+            type: 'string'
+        )
+    )]
+    #[OA\Response(
+        response: Response::HTTP_BAD_REQUEST,
+        description: 'Bad request',
+        content: new OA\JsonContent(
+            type: 'string'
+        )
     )]
     #[OA\Tag(name: 'Bonuses')]
-    public function __invoke(Bonus $bonus, BaseManager $manager): Response
-    {
-        $manager->delete($bonus);
+    public function __invoke(
+        int $bonusId,
+        BaseManager $manager,
+        BonusRepository $bonuses
+    ): Response {
+        if ($bonus = $bonuses->find($bonusId)) {
+            $manager->delete($bonus);
 
-        return $this->json('Bonus was deleted successfully');
+            $message = 'Bonus was deleted successfully';
+            $status = Response::HTTP_OK;
+        } else {
+            $message = "Bonus $bonusId not found";
+            $status = Response::HTTP_BAD_REQUEST;
+        }
+
+        return $this->json($message, $status);
     }
 }
